# Save a string allocation in visit_STAR