@eugeneius 2020-06-30
The benchmark above calls `visit_STAR` with a `Cat` node, which isn't possible in a real application. When is `visit_STAR` actually called? When does `@matchers[node.left.to_sym]` return nil?

I understand that this optimisation is local to the method and doesn't depend on the above details, but it's hard to be excited about merging a change that saves one object allocation with no context on how often the code being optimised runs.