@vinistock 2020-07-08
Indeed the node was incorrect. I have remade the benchmark using a star node.

Basically, when visiting the different nodes it will try to find a defined matcher for the anchored regexp or else return the string literal. Matchers is just a hash, so you define the star + whatever matcher name you want. For example, if the node is `*format` it will try to find a matcher named `format`.

All visit methods are invoked with meta programming. The generic method `visit` will check the type of the node being visited and then dispatch the right method.

In summary, these changes save an allocation whenever we have `*something` and `something` isn't a defined matcher. Matchers are extracted from the constraints defined for routes.

Here's the new benchmark and result.

**Benchmark**
```ruby
# frozen_string_literal: true

require "bundler/inline"

gemfile(true) do
  source "https://rubygems.org"

  git_source(:github) { |repo| "https://github.com/#{repo}.git" }

  gem "rails", path: "../rails", require: "rails/all"
  gem "benchmark-ips"
  gem "benchmark-memory", require: "benchmark/memory"
end

require "action_dispatch/journey/path/pattern"

module ActionDispatch
  module Journey # :nodoc:
    module Path # :nodoc:
      class Pattern # :nodoc:
        class AnchoredRegexp < Journey::Visitors::Visitor
          def fast_visit_STAR(node)
            re = @matchers[node.left.to_sym]
            re ? "(#{re})" : "(.+)"
          end
        end
      end
    end
  end
end

ast = ActionDispatch::Journey::Parser.parse "*format"

SCENARIOS = {
  "WITH A MATCHER" => { format: /.+/ },
  "WITHOUT A MATCHER" => {}
}

SCENARIOS.each_pair do |name, value|
  puts "*" * 80
  puts name
  puts "*" * 80

  anchored = ActionDispatch::Journey::Path::Pattern::AnchoredRegexp.new(%w( / . ? ).join, value)

  Benchmark.ips do |x|
    x.report("visit_STAR")      { anchored.visit_STAR(ast) }
    x.report("fast_visit_STAR") { anchored.fast_visit_STAR(ast) }
    x.compare!
  end

  Benchmark.memory do |x|
    x.report("visit_STAR")      { anchored.visit_STAR(ast) }
    x.report("fast_visit_STAR") { anchored.fast_visit_STAR(ast) }
    x.compare!
  end
end
```
**Results**
```
********************************************************************************
WITH A MATCHER
********************************************************************************
Warming up --------------------------------------
          visit_STAR   158.880k i/100ms
     fast_visit_STAR   163.316k i/100ms
Calculating -------------------------------------
          visit_STAR      1.550M (± 6.4%) i/s -      7.785M in   5.041130s
     fast_visit_STAR      1.545M (± 7.4%) i/s -      7.839M in   5.098183s

Comparison:
          visit_STAR:  1550101.8 i/s
     fast_visit_STAR:  1545032.7 i/s - same-ish: difference falls within error

Calculating -------------------------------------
          visit_STAR   144.000  memsize (     0.000  retained)
                         2.000  objects (     0.000  retained)
                         2.000  strings (     0.000  retained)
     fast_visit_STAR   144.000  memsize (    40.000  retained)
                         2.000  objects (     1.000  retained)
                         2.000  strings (     1.000  retained)

Comparison:
          visit_STAR:        144 allocated
     fast_visit_STAR:        144 allocated - same
********************************************************************************
WITHOUT A MATCHER
********************************************************************************
Warming up --------------------------------------
          visit_STAR   362.288k i/100ms
     fast_visit_STAR   620.486k i/100ms
Calculating -------------------------------------
          visit_STAR      3.592M (± 3.5%) i/s -     18.114M in   5.049222s
     fast_visit_STAR      6.063M (± 3.0%) i/s -     30.404M in   5.019259s

Comparison:
     fast_visit_STAR:  6062982.4 i/s
          visit_STAR:  3591997.1 i/s - 1.69x  (± 0.00) slower

Calculating -------------------------------------
          visit_STAR    40.000  memsize (    40.000  retained)
                         1.000  objects (     1.000  retained)
                         1.000  strings (     1.000  retained)
     fast_visit_STAR     0.000  memsize (     0.000  retained)
                         0.000  objects (     0.000  retained)
                         0.000  strings (     0.000  retained)

Comparison:
     fast_visit_STAR:          0 allocated
          visit_STAR:         40 allocated - Infx more
```