@eugeneius 2020-07-08
Thanks for the explanation, this is what I was missing:

> Matchers are extracted from the constraints defined for routes.

Most routes don't have constraints, so `@matchers` is typically empty, and the extra allocation is the most common case.